Abstract Renderer2D

+++ Blanco/src/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -12,7 +12,7 @@ namespace Blanco
 
 	void OpenGLRendererAPI::SetViewport(float x, float y, float width, float height)
 	{
-		glViewport(x, y, width, height);
+		glViewport((int)x, (int)y, (int)width, (int)height);
 	}
 
 	void OpenGLRendererAPI::SetClearColor(const glm::vec4& color)
